feat(characters): continue
