+++ src/lib/locales/ru/admin.json
@@ -3,6 +3,8 @@
     "subtitle": "Панель управления: Галерея",
     "hint": "Выберите портрет для редактирования. Установите флажки, чтобы выбрать несколько портретов",
     "filter": "Отфильтровать",
+    "characters": "Персонажи",
+    "portraits": "Портреты",
     "upload": "Загрузить изображения",
     "statistics": "Статистика"
   },
