+++ src/routes/(user)/+layout.svelte
@@ -1,1 +1,5 @@
+<svelte:head>
+  <title>Deorum</title>
+</svelte:head>
+
 <slot />
