+++ src/routes/admin/portraits/statistics/colors/+page.svelte
