+++ src/routes/admin/portraits/statistics/originals/+page.svelte
