+++ src/routes/admin/portraits/statistics/quality/+page.svelte
