+++ src/routes/admin/portraits/table/+page.ts
@@ -1,0 +1,6 @@
+import { redirect } from '@sveltejs/kit';
+
+/** @type {import('./$types').PageLoad} */
+export async function load() {
+  throw redirect(308, '/admin/portraits/1');
+}
