+++ src/routes/admin/portraits/table/[slug]/+page.svelte
